Add the first 3 models, and seed Users
There is still a lot wrong with this commit. For some reason the
model definitions/migrations don't result in Postgres tables
with the proper default and on-update values, which means that
the UUID field, as well as createdAt and updatedAt, both create
errors when running the seed, and will on other actions too.

For the time being I'm getting around this just by generating
the values in code (via node's uuid package) and by manually
setting the field defaults in Postgres. This is all a big @TODO.